created geocode router file and moved geocode route
